go-mysqlx, go-binlog: regenerate code using proto 0.5

+++ generate-go.sh
@@ -2,9 +2,18 @@ cd proto
 
 echo "generate go code..."
 GOGO_ROOT=${GOPATH}/src/github.com/gogo/protobuf
-protoc -I.:${GOGO_ROOT}:${GOGO_ROOT}/protobuf --gofast_out=../go-tipb *.proto
-cd ../go-tipb
-sed -i.bak -E 's/import _ \"gogoproto\"//g' *.pb.go
-sed -i.bak -E 's/import fmt \"fmt\"//g' *.pb.go
-rm -f *.bak
-goimports -w *.pb.go
+if [ ! -d $GOGO_ROOT ]; then
+	echo "go get -u github.com/gogo/protobuf/protoc-gen-gofast"
+	echo "cd /Users/zhexuany/repo/go/src/github.com/gogo/protobuf"
+	echo "git checkout v0.5"
+	echo "rm /Users/zhexuany/repo/go/bin/protoc-gen-gofast"
+	echo "go get github.com/gogo/protobuf/protoc-gen-gofast"
+else
+	protoc -I.:${GOGO_ROOT}:${GOGO_ROOT}/protobuf --gofast_out=../go-tipb *.proto
+	cd ../go-tipb
+	sed -i.bak -E 's/import _ \"gogoproto\"//g' *.pb.go
+	sed -i.bak -E 's/import fmt \"fmt\"//g' *.pb.go
+	rm -f *.bak
+	goimports -w *.pb.go
+fi
+
